+++ packages/Filter/src/helpers/testers.js
@@ -17,6 +17,14 @@ const processDate = (momentParsingFormat, value, testValue, testFunction) => {
   return false;
 };
 
+const isOneOf = (rowValue, multiSelectSelectedValues) => {
+  return multiSelectSelectedValues.length === 0 || multiSelectSelectedValues.includes(rowValue);
+};
+
+const isNotOneOf = (rowValue, multiSelectSelectedValues) => {
+  return multiSelectSelectedValues.length === 0 || !multiSelectSelectedValues.includes(rowValue);
+};
+
 const testers = {
   [Filter.rules.IS]: is,
   [Filter.rules.IS_NOT]: isNot,
@@ -38,6 +46,8 @@ const testers = {
     processDate(momentParsingFormat, value, testValue, (a, b) => a.isBefore(b)),
   [Filter.rules.IS_AFTER]: (value, testValue, { momentParsingFormat }) =>
     processDate(momentParsingFormat, value, testValue, (a, b) => a.isAfter(b)),
+  [Filter.rules.IS_ONE_OF]: isOneOf,
+  [Filter.rules.IS_NOT_ONE_OF]: isNotOneOf,
 };
 
 export default testers;