+++ packages/Textarea/stories/examples/Showcase.js
@@ -1,30 +1,24 @@
 import React from "react";
 import { boolean, select, text } from "@storybook/addon-knobs";
-import { Rule, Story, Tagline } from "storybook/assets/styles/common.styles";
-import Heading from "@paprika/heading";
-import * as types from "../../src/types";
-import TextareaExample from "./TextareaExample";
+import CodeViewer from "storybook/components/CodeViewer";
+import Textarea from "../../src";
 
-const textareaProps = () => ({
-  a11yText: text("a11yText"),
+const getKnobs = () => ({
+  size: select("size", Object.values(Textarea.types.size), Textarea.types.size.MEDIUM),
   canExpand: boolean("canExpand", true),
+  maxHeight: text("maxHeight", 300),
+  minHeight: text("minHeight", 80),
   hasError: boolean("hasError", false),
   isDisabled: boolean("isDisabled", false),
   isReadOnly: boolean("isReadOnly", false),
-  maxHeight: text("maxHeight"),
-  placeholder: text("placeholder", "This is a default placeholder..."),
-  size: select("size", [types.SMALL, types.MEDIUM, types.LARGE], types.MEDIUM),
+  placeholder: text("placeholder", ""),
+  a11yText: text("a11yText"),
 });
 
-const ExampleStory = props => (
-  <Story>
-    <Heading level={1} displayLevel={2} isLight>
-      Showcase
-    </Heading>
-    <Tagline>Use the knobs to tinker with the props.</Tagline>
-    <Rule />
-    <TextareaExample {...props} />
-  </Story>
-);
-
-export default () => <ExampleStory {...textareaProps()} />;
+export default function Showcase() {
+  return (
+    <CodeViewer>
+      <Textarea {...getKnobs()} />
+    </CodeViewer>
+  );
+}